Add type casting to print properly

+++ core/lz4/src/ZipLZ4.cxx
@@ -170,7 +170,7 @@ void R__zipLZ4BS(int cxlevel, int *srcsize, char *src, int *tgtsize, char *tgt, 
       R__zipLZ4(cxlevel, srcsize, src, tgtsize, tgt, irep);
       return;
    } else if (returnStatus_lz4bs < 0) {
-      fprintf(stderr, "Bitshuffle failed: got negative returnStatus %lld\n", returnStatus_lz4bs);
+      fprintf(stderr, "Bitshuffle failed: got negative returnStatus %lld\n", (long long) returnStatus_lz4bs);
       return;
    }
    int64_t returnStatus = bshuf_compress_lz4(src, &tgt[kHeaderSize], elem_count, sizeof(float), 0);
@@ -225,7 +225,7 @@ void R__unzipLZ4BS(int * srcsize, unsigned char * src, int * tgtsize, unsigned c
         fprintf(
            stderr,
            "R__unzipLZ4: Buffer corruption error!  Calculated checksum %llu; checksum calculated in the file was %llu.\n",
-           checksumResult, checksumFromFile);
+           (unsigned long long) checksumResult, (unsigned long long) checksumFromFile);
         return;
      } 
      size_t elem_count = *tgtsize / sizeof(float);
